+++ config/updates.ts
@@ -25,18 +25,18 @@ import { registerAs } from '@nestjs/config';
 export default registerAs('updates', () => ({
   api: {
     versionURL: process.env.API_VERSION_URL || 'https://raw.githubusercontent.com/soltecsis/fwcloud-api',
-    installDir: process.env.API_INSTALL_DIR || '/opt/fwcloud-api'
+    installDir: process.env.API_INSTALL_DIR || '/opt/fwcloud/api'
   },
   ui: {
     versionURL: process.env.UI_VERSION_URL || 'https://raw.githubusercontent.com/soltecsis/fwcloud-ui',
-    installDir: process.env.UI_INSTALL_DIR || '/opt/fwcloud-ui'
+    installDir: process.env.UI_INSTALL_DIR || '/opt/fwcloud/ui'
   },
   updater: {
     versionURL: process.env.UPDATER_VERSION_URL || 'https://raw.githubusercontent.com/soltecsis/fwcloud-updater',
-    installDir: process.env.UPDATER_INSTALL_DIR || '/opt/fwcloud-updater'
+    installDir: process.env.UPDATER_INSTALL_DIR || '/opt/fwcloud/updater'
   },
   websrv: {
     versionURL: process.env.WEBSRV_VERSION_URL || 'https://raw.githubusercontent.com/soltecsis/fwcloud-websrv',
-    installDir: process.env.WEBSRV_INSTALL_DIR || '/opt/fwcloud-websrv'
+    installDir: process.env.WEBSRV_INSTALL_DIR || '/opt/fwcloud/websrv'
   }
 }));